requirements.txt
Flask>=2.2
Flask-SQLAlchemy==3.0.2
